webui: hide development form in non-development environment.

+++ package/thingino-webui/files/var/www/x/_common.cgi
@@ -483,6 +483,9 @@ menu() {
 			grep -q -s "^${p}_enabled=\"true\"" $ui_config_dir/${p}.conf && echo -n " plugin-enabled"
 			echo "\" href=\"$i\">$n</a></li>"
 		else
+			# FIXME: dirty hack
+			[ "$i" = "config-developer.cgi" ] && [ ! -f /etc/init.d/S44devmounts ] && continue
+
 			n="$(sed -r -n '/page_title=/s/^.*page_title="(.*)",*$/\1/p' $i)"
 			echo -n "<li><a class=\"dropdown-item\" href=\"$i\">$n</a></li>"
 		fi
